Fix ignored tests

+++ src/main/java/org/rutebanken/tiamat/repository/PathLinkRepositoryImpl.java
@@ -23,7 +23,6 @@ import jakarta.persistence.Query;
 import org.apache.commons.lang3.StringUtils;
 import org.rutebanken.tiamat.model.PathLink;
 
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
@@ -50,11 +49,11 @@ public class PathLinkRepositoryImpl implements PathLinkRepositoryCustom {
 
         try {
             @SuppressWarnings("unchecked")
-            List<BigInteger> results = query.getResultList();
+            List<Long> results = query.getResultList();
             if (results.isEmpty()) {
                 return null;
             } else {
-                return results.getFirst().longValue();
+                return results.getFirst();
             }
         } catch (NoResultException noResultException) {
             return null;
